room add

+++ admin/ajax/features_facilities.php
@@ -35,9 +35,17 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values, 'i');
-        echo $res;
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?', [$frm_data['rem_feature']], 'i');
+
+        if(mysqli_num_rows($check_q)==0){
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values, 'i');
+            echo $res;
+        }else{
+            echo 'room_added';
+        }
+
+       
     }
     
     if (isset($_POST['action']) && $_POST['action'] == 'add_facility') {
@@ -96,7 +104,10 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facility']];
 
-        $pre_q = "SELECT * FROM `facilities` WHERE `id`= ?";
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?', [$frm_data['rem_facility']], 'i');
+
+        if(mysqli_num_rows($check_q)==0){
+            $pre_q = "SELECT * FROM `facilities` WHERE `id`= ?";
         $res = select($pre_q, $values, 'i');
         $img = mysqli_fetch_assoc($res);
     
@@ -107,5 +118,10 @@
         } else {
             echo 0;
         }
+        }else{
+            echo 'room_added';
+        }
+
+        
     }
 ?>
